Football Results Generator
This takes in football or any results and creates a report based on the
results inputted.

input.txt is an example file with valid and invalid entries in it.

Make sure that the HTML directory is created or another is used.

Entries into the console must be made in the format:

HomeTeam : AwayTeam : HomeScore : AwayScore